Added netty-codec to the dependencies, although the code using it does not appear to fire properly.
Configured 10 child threads to start with on the server.

Added a ChannelHandlerContext.close() to close incoming TCP requests.  Also,
we're subclassing ByteToMessageDecoder instead of ChannelInboundHandlerAdapter
but the method that adds, "decode", does not appear to be firing.  Also, load
testing indicates there might be issues with messages stepping on each other.